django/pyproject.toml
[tool.poetry]
name = "Thunderstore"
version = "0.0.0"
description = "Thunderstore is a mod database and API for downloading mods"
authors = ["Mythic <mythicmaniac@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1"
django-environ = "^0.4"
arrow = "^0.17"
Pillow = "^7.1"
psycopg2-binary = "^2.8"
easy-thumbnails = "^2.7"  # This has a monkeypatch and might be a difficult upgrade
gunicorn = { version = "^20.0", extras = ["gevent"] }
social-auth-app-django = "^4.0"
social-auth-core = "^4.0"
whitenoise = "^5.0"
django-storages = { version = "^1.11", extras = ["google", "boto3"] }
djangorestframework = "^3.11"
drf-yasg = "^1.17"
django-ipware = "^3.0"
sentry_sdk = "^0.19"
django-redis = "^4.11"
pyjwt = { version = "^2.0", extras = ["crypto"] }
celery = "^5.0.5"
django-celery-results = "^2.0.0"
django-celery-beat = "^2.1.0"
ulid2 = "^0.2.0"
django-cachalot = "^2.3.3"
django-cors-headers = "^3.7.0"
boto3 = "^1.17.47"
bleach = "^3.3.0"
markdown-it-py = {extras = ["linkify"], version = "^1.1.0"}

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^2.8"
pytest-django = "^4.1"
factory_boy = "^3.2"
Faker = "^6.0"
pytest-mock = "^3.5"
django-debug-toolbar = "^3.2"
pytest-env = "^0.6.2"
black = ">=20.8b1"
isort = "^5.7.0"
pre-commit = "^2.9.3"
mypy = ">=0.790"
django-stubs = "^1.7.0"
djangorestframework-stubs = "^1.3.0"

flake8 = "^3.8.2"
flake8-2020 = "^1.6.0"
flake8-broken-line = "^0.3.0"
flake8-bugbear = ">=20.11.1"
flake8-builtins = "^1.5.3"
flake8-commas = "^2.0.0"
flake8-comprehensions = "^3.3.1"
flake8-pie = "^0.6.1"
flake8-printf-formatting = "^1.1.0"
flake8-pytest-style = "^1.3.0"
pep8-naming = "^0.11.1"
watchdog = {extras = ["watchmedo"], version = "^1.0.2"}
mypy-boto3-s3 = "^1.17.47"

[tool.black]
target-version = ["py38"]
exclude = '''
(
  /(
      \.git
    | \.mypy_cache
    | \.venv
    | \.vscode
    | \.idea
  )/
)
'''

[tool.isort]
profile = "black"
known_third_party = ["django"]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
